backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
redis
celery
slowapi
requests
python-jose
bcrypt>=4.0.0
pydantic>=1.10
pydantic-settings>=2
email-validator
python-multipart

